remove the -tune option, and consolidated the config files
