--- python/perf-kernels/fused_moe/configs/E=8,N=14336,device_name=AMD_Instinct_MI300X.json
@@ -1,200 +1,0 @@
-{
-  "1": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 32,
-    "BLOCK_SIZE_K": 256,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 2,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 1
-  },
-  "2": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 16,
-    "BLOCK_SIZE_K": 128,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 2,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "4": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 32,
-    "BLOCK_SIZE_K": 256,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 2,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "8": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 16,
-    "BLOCK_SIZE_K": 256,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 1,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "16": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 16,
-    "BLOCK_SIZE_K": 256,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 4,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "24": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 32,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 1,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "32": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 16,
-    "BLOCK_SIZE_K": 128,
-    "GROUP_SIZE_M": 4,
-    "num_warps": 2,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 1
-  },
-  "48": {
-    "BLOCK_SIZE_M": 16,
-    "BLOCK_SIZE_N": 16,
-    "BLOCK_SIZE_K": 128,
-    "GROUP_SIZE_M": 4,
-    "num_warps": 2,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "64": {
-    "BLOCK_SIZE_M": 32,
-    "BLOCK_SIZE_N": 64,
-    "BLOCK_SIZE_K": 128,
-    "GROUP_SIZE_M": 4,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "96": {
-    "BLOCK_SIZE_M": 32,
-    "BLOCK_SIZE_N": 32,
-    "BLOCK_SIZE_K": 128,
-    "GROUP_SIZE_M": 4,
-    "num_warps": 4,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "128": {
-    "BLOCK_SIZE_M": 64,
-    "BLOCK_SIZE_N": 64,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 4,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "256": {
-    "BLOCK_SIZE_M": 128,
-    "BLOCK_SIZE_N": 128,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 4,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 1
-  },
-  "512": {
-    "BLOCK_SIZE_M": 128,
-    "BLOCK_SIZE_N": 128,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 4,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "1024": {
-    "BLOCK_SIZE_M": 256,
-    "BLOCK_SIZE_N": 256,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "1536": {
-    "BLOCK_SIZE_M": 128,
-    "BLOCK_SIZE_N": 128,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "2048": {
-    "BLOCK_SIZE_M": 128,
-    "BLOCK_SIZE_N": 128,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  },
-  "3072": {
-    "BLOCK_SIZE_M": 128,
-    "BLOCK_SIZE_N": 128,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 1
-  },
-  "4096": {
-    "BLOCK_SIZE_M": 256,
-    "BLOCK_SIZE_N": 256,
-    "BLOCK_SIZE_K": 64,
-    "GROUP_SIZE_M": 1,
-    "num_warps": 8,
-    "num_stages": 2,
-    "waves_per_eu": 0,
-    "matrix_instr_nonkdim": 16,
-    "kpack": 2
-  }
-}
